Adding letter tracker
Have added a letter tracking feature to capture instances where there are two of the same letter in the guess word and only one is in the correct position. Known issue: this only works when the correct position is before the incorrect one, and this will need to be fixed to cover the other instance. I have also replaced my code to match the indexes with the inbuilt enumerate function.